🧹 Polish: Extract canvas drawing helpers from QRCanvas
- Extracted `drawRoundRect`, `drawPoly`, `drawStar`, `drawRoughRect`, and `drawScribble` to `src/utils/canvasHelpers.ts`.
- Refactored `QRCanvas.tsx` to import these helpers.
- Standardized indentation in the new utility file to 2 spaces.
- Verified with full test suite - NO LOGIC CHANGES.
